feat: enhance expression support with generated methods and typing improvements
- Add generated methods for `Build`, `ToExpr`, and `As` across `TimeExpr`, `DateExpr`, `YearExpr`, and related types.
- Extend SQL function support, including `Sum`, `Avg`, `Round`, and `Truncate`, for numeric and temporal expressions.
- Adjust `Ceil`, `Floor`, and `Pow` to better align with expected return types for `DecimalExpr` and `IntExpr`.
- Introduce `@gentype` tags and `@gen` annotations to streamline code generation for expressions.
- Remove redundant method definitions and consolidate shared implementations for arithmetic and null-handling logic.

+++ internal/fields/decimal_test.go
@@ -134,7 +134,7 @@ func TestDecimalExprT_Ceil(t *testing.T) {
 	expr := NewDecimalExpr[float64](clause.Expr{SQL: "price", Vars: nil})
 	result := expr.Ceil()
 
-	e := result.Gte(int64(0))
+	e := result.Gte(float64(0))
 	_, ok := e.(clause.Expr)
 	assert.True(t, ok)
 }
@@ -143,7 +143,7 @@ func TestDecimalExprT_Floor(t *testing.T) {
 	expr := NewDecimalExpr[float64](clause.Expr{SQL: "price", Vars: nil})
 	result := expr.Floor()
 
-	e := result.Gte(int64(0))
+	e := result.Gte(float64(0))
 	_, ok := e.(clause.Expr)
 	assert.True(t, ok)
 }
